Added Documentation and Implemented PathfinderController
